requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
einops==0.6.1
fairscale==0.4.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
google-auth==2.19.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
hydra-core==1.0.7
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
infinibatch==0.1.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.23.0
oauthlib==3.2.2
omegaconf==2.0.6
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pip==23.0.1
# protobuf==4.23.2
protobuf==3.20.0 
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.4.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
setuptools==67.8.0
six==1.16.0
sympy==1.12
tiktoken==0.4.0
timm==0.4.12
torchmetrics==0.5.0
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.2
urllib3==1.26.16
Werkzeug==2.3.4
wheel==0.38.4
yarl==1.9.2
zipp==3.15.0
